requirements.txt
# Inspiration Bot - Dependencies

# Telegram
python-telegram-bot>=20.7

# Google Gemini AI
google-generativeai>=0.8.0

# Scheduler
apscheduler>=3.10.0

# Settings
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP Server (Railway Health Check)
aiohttp>=3.9.1

# Utilities
pytz>=2023.3
python-dotenv>=1.0.0
loguru>=0.7.2
